+++ examples/src/CASCADE_DEMO/main.cpp
@@ -27,7 +27,11 @@ int main(int argc, char **argv)
 
 	// Init the bcm2835 library
 	printf("Test Begin :: MAX7219_7SEG_RPI\r\n");
-	if(!bcm2835_init()) {return -1;}
+		if(!bcm2835_init())  // Init the bcm2835 library
+	{
+		printf("Error 1201 :: bcm2835_init failed. Are you running as root??\n");
+		return -1;
+	}
 	
 	// Some test data defined
 	char teststr1[] = "Display1";
@@ -36,11 +40,19 @@ int main(int argc, char **argv)
 	// Init the displays
 	// -Init Display one 
 	myMAX.SetCurrentDisplayNumber(1);
-	myMAX.InitDisplay(myMAX.ScanEightDigit, myMAX.DecodeModeNone);
+	if(!myMAX.InitDisplay(myMAX.ScanEightDigit, myMAX.DecodeModeNone))
+	{
+		printf("Error 1202 :: bcm2835_spi_begin failed. Are you running as root??\n");
+		return -1;
+	}
 	myMAX.ClearDisplay();
 	// -Init Display Two
 	myMAX.SetCurrentDisplayNumber(2);
-	myMAX.InitDisplay(myMAX.ScanEightDigit, myMAX.DecodeModeNone);
+	if(!myMAX.InitDisplay(myMAX.ScanEightDigit, myMAX.DecodeModeNone))
+	{
+		printf("Error 1202 :: bcm2835_spi_begin failed. Are you running as root??\n");
+		return -1;
+	}
 	myMAX.ClearDisplay();
 
 	// Write to Display one