+++ examples/src/TESTS/main.cpp
@@ -36,6 +36,9 @@
 // Constructor object 
 MAX7219_SS_RPI myMAX(CLK, CS ,DIN);
 
+// Function Prototypes
+bool Setup(void);
+void EndTest(void);
 void Test1(void);
 void Test2(void);
 void Test3(void);
@@ -53,11 +56,7 @@ void Test11(void);
 int main(int argc, char **argv) 
 {
 
-	// Init the bcm2835 library
-	printf("Test Begin :: MAX7219_7SEG_RPI\r\n");
-	if(!bcm2835_init()) {return -1;}
-
-	myMAX.InitDisplay(myMAX.ScanEightDigit, myMAX.DecodeModeNone);
+	if (!Setup()) return -1;
 	
 	Test1();
 	Test2();
@@ -70,14 +69,37 @@ int main(int argc, char **argv)
 	Test9();
 	Test10();
 	
-	myMAX.DisplayEndOperations();
-	
-	// Close the bcm2835  library
-	bcm2835_close(); 
-	printf("Test End\r\n");
+	EndTest();
 	return 0;
 } 
 // End of main
+
+// Functions
+
+// Setup test
+bool Setup(void)
+{
+	printf("Test Begin :: MAX7219_7SEG_RPI\r\n");
+	if(!bcm2835_init())  // Init the bcm2835 library
+	{
+		printf("Error 1201 :: bcm2835_init failed. Are you running as root??\n");
+		return false;
+	}
+	printf("bcm2835 library Version Number :: %u\r\n",bcm2835_version());
+	printf("MAX7219_7SEG Library version number :: %u\r\n", myMAX.GetLibVersionNum()); 
+	myMAX.InitDisplay(myMAX.ScanEightDigit, myMAX.DecodeModeNone);
+	myMAX.ClearDisplay();
+	return true;
+}
+
+// Clean up before exit
+void EndTest(void)
+{
+	myMAX.DisplayEndOperations();
+	bcm2835_close();  // Close the bcm2835 library
+	printf("Test End\r\n");
+}
+
 
 void Test1(void)
 {