+++ extra/doc/CHANGELOG.md
@@ -2,4 +2,7 @@
 
 * version 1.2.0 May 2023
 	* first release
+* version 1.3.0 Dec 2023
+	* Changed Makefile CCFLAGS so it complies on 64 bit system without errors.
+	* Changed "InitDisplay" function so it returns false if SPi-begin fails This catches user error of not running as root.